refactor client master typing
